Move overlay function to common location.
